move some admin components to components directory
