+++ src/scenes/AdminControl/AdminControl.js
@@ -6,8 +6,8 @@ import AdminAddLocation from "./Components/AdminAddLocation";
 import AdminAddDeliverableType from "./Components/AdminAddDeliverableType";
 import AdminAddRiderResponsibility from "./Components/AdminAddRiderResponsibility";
 import { Stack } from "@mui/material";
-import { RiderResponsibilityChips } from "./RiderResponsibilityChips";
-import { DeliverableTypeChips } from "./DeliverableTypeChips";
+import { RiderResponsibilityChips } from "./Components/RiderResponsibilityChips";
+import { DeliverableTypeChips } from "./Components/DeliverableTypeChips";
 
 export function AdminControl() {
     return (
